requirements.txt
aiogram==3.13.0
sqlalchemy==2.0.35
apscheduler==3.10.4
python-dotenv==1.0.1
